// packages/data-atlas-wizard/src/generated/DataChain.ts
/*
 * This file was generated by emfts-codegen.
 * DO NOT EDIT - changes will be overwritten.
 *
 * @generated
 */

import type { EObject, BasicEObject, EClass, EStructuralFeature } from '@emfts/core';
import type { DataSourceConfig } from './DataSourceConfig';
import type { DatasetConfig } from './DatasetConfig';
import type { ExportConfig } from './ExportConfig';

/**
 * DataChain
 * @generated
 */
export interface DataChain extends EObject {
  id: string;
  source?: DataSourceConfig;
  sharedSource?: DataSourceConfig;
  datasets: DatasetConfig[];
  exports: ExportConfig[];

}