// packages/data-atlas-wizard/src/generated/ExportConfig.ts
/*
 * This file was generated by emfts-codegen.
 * DO NOT EDIT - changes will be overwritten.
 *
 * @generated
 */

import type { EObject, BasicEObject, EClass, EStructuralFeature } from '@emfts/core';
import type { ExportKind } from './ExportKind';

/**
 * ExportConfig
 * @generated
 */
export interface ExportConfig extends EObject {
  selected?: boolean;
  kind: ExportKind;
  id: string;
  name: string;
  description: string;
  separator?: string;
  includeTypeHeader?: boolean;

}